+++ pm4py_ucm/__init__.py
@@ -66,4 +66,4 @@ __all__ = [
     "annotate_performance",
 ]
 
-__version__ = "0.6.0"
+__version__ = "0.5.1"